refactor: move millisecondsToMmSs to dates utils
